perf(nuxt): split head runtime plugin by environment
Replace the single universal `nuxt:head` runtime plugin with environment-
specific plugins so Vite tree-shakes per environment:

- `unhead.server.ts` installs `ssrContext.head` (+ island head freeze) and
  no longer pulls the client head factories into the SSR bundle.
- `unhead.client.ts` uses the non-streaming `createClientHead` path.
- `unhead-stream.client.ts` is registered only when SSR streaming is
  enabled; the non-streaming client build no longer imports
  `@unhead/vue/stream/client`.

The client variant is chosen at build time via `ssrStreamingEnabled`
instead of the runtime `ssrStreaming` flag, removing the unused streaming
factory from the default client bundle. The now-unused universal
`runtime/plugins/unhead.ts` and the orphaned `ssrStreaming` export from the
`unhead.config.mjs` template are removed. The shared client head wiring
(DOM pause/resume, deferred entry disposal) lives in `install-client-head.ts`
so the two client variants stay in sync. Behaviour is preserved.
