--- packages/nuxt/src/head/runtime/plugins/unhead.ts
@@ -1,75 +1,0 @@
-import { createHead as createClientHead } from '@unhead/vue/client'
-import { createStreamableHead as createStreamableClientHead } from '@unhead/vue/stream/client'
-import type { ActiveHeadEntry } from '@unhead/vue'
-import { defineNuxtPlugin } from '#app/nuxt'
-import type { ObjectPlugin, Plugin } from '#app/nuxt'
-import { freezeHead } from '../island-head'
-
-// @ts-expect-error virtual file
-import unheadOptions from '#build/unhead-options.mjs'
-// @ts-expect-error virtual file
-import { ssrStreaming } from '#build/unhead.config.mjs'
-
-const plugin: Plugin & ObjectPlugin = defineNuxtPlugin({
-  name: 'nuxt:head',
-  enforce: 'pre',
-  setup (nuxtApp) {
-    // `createStreamableHead` consumes the `window.__unhead__` queue
-    // populated during SSR streaming, falling back to `createClientHead`
-    // when no queue is found (e.g. bot requests served buffered).
-    const head = import.meta.server
-      ? nuxtApp.ssrContext!.head
-      : ssrStreaming
-        ? (createStreamableClientHead(unheadOptions) || createClientHead(unheadOptions))
-        : createClientHead(unheadOptions)
-
-    // Drop plugin-phase `useHead` writes for islands -- they belong to the
-    // surrounding route, not the island response. Unfreeze on `app:created`
-    // (after `applyPlugins` resolves) so island components write normally.
-    if (import.meta.server && nuxtApp.ssrContext!.islandContext) {
-      const unfreeze = freezeHead(head)
-      nuxtApp.hooks.hookOnce('app:created', unfreeze)
-    }
-
-    // nuxt.config appHead is set server-side within the renderer
-    nuxtApp.vueApp.use(head)
-
-    if (import.meta.client) {
-      // pause dom updates until page is ready and between page transitions
-      let pauseDOMUpdates = true
-      const syncHead = () => {
-        pauseDOMUpdates = false
-        head.render()
-      }
-      head.hooks?.hook('dom:beforeRender', (context) => { context.shouldRender = !pauseDOMUpdates })
-      nuxtApp.hooks.hook('page:start', () => { pauseDOMUpdates = true })
-      // wait for new page before unpausing dom updates (triggered after suspense resolved)
-      nuxtApp.hooks.hook('page:finish', () => {
-        // app:suspense:resolve hook will unpause the DOM
-        if (!nuxtApp.isHydrating) { syncHead() }
-      })
-      // unpause on error
-      nuxtApp.hooks.hook('app:error', syncHead)
-      // unpause the DOM once the mount suspense is resolved
-      nuxtApp.hooks.hook('app:suspense:resolve', syncHead)
-
-      // Defer head-entry disposal during a page transition.
-      const originalPush = head.push.bind(head)
-      head.push = ((input: Parameters<typeof head.push>[0], options?: Parameters<typeof head.push>[1]) => {
-        const entry = originalPush(input, options) as ActiveHeadEntry<typeof input>
-        const originalDispose = entry.dispose.bind(entry)
-        entry.dispose = () => {
-          const transitionPromise = nuxtApp['~transitionPromise']
-          if (transitionPromise) {
-            transitionPromise.then(originalDispose)
-          } else {
-            originalDispose()
-          }
-        }
-        return entry
-      }) as typeof head.push
-    }
-  },
-})
-
-export default plugin
