+++ packages/nuxt/test/app.test.ts
@@ -69,8 +69,12 @@ describe('resolveApp', () => {
             "src": "<rootDir>/.nuxt/components.plugin.mjs",
           },
           {
-            "mode": "all",
-            "src": "<repoRoot>/packages/nuxt/src/head/runtime/plugins/unhead.ts",
+            "mode": "server",
+            "src": "<repoRoot>/packages/nuxt/src/head/runtime/plugins/unhead.server.ts",
+          },
+          {
+            "mode": "client",
+            "src": "<repoRoot>/packages/nuxt/src/head/runtime/plugins/unhead.client.ts",
           },
           {
             "mode": "all",
@@ -80,6 +84,24 @@ describe('resolveApp', () => {
         "rootComponent": "<repoRoot>/packages/nuxt/src/app/components/nuxt-root.vue",
         "templates": [],
       }
+    `)
+  })
+
+  it('registers the streaming client head plugin when ssrStreaming is enabled', async () => {
+    const app = await getResolvedApp([
+      {
+        name: 'nuxt.config.ts',
+        contents: 'export default defineNuxtConfig({ experimental: { ssrStreaming: true } })',
+      },
+    ])
+    const headPlugins = app.plugins
+      .filter(p => !('getContents' in p) && p.src.includes('head/runtime/plugins'))
+      .map(p => p.src)
+    expect(headPlugins).toMatchInlineSnapshot(`
+      [
+        "<repoRoot>/packages/nuxt/src/head/runtime/plugins/unhead.server.ts",
+        "<repoRoot>/packages/nuxt/src/head/runtime/plugins/unhead-stream.client.ts",
+      ]
     `)
   })
